No change. Probably uses the same type behind the scenes

+++ include/Hanoi.h
@@ -6,7 +6,7 @@
 
 class Hanoi {
     public:
-        Hanoi(std::uint64_t numDisks);
+        Hanoi(std::uint_fast8_t numDisks);
 
         void solve();
 
@@ -14,15 +14,15 @@ class Hanoi {
 
         void print();
 
-        bool swap(std::stack<std::uint64_t, std::vector<std::uint64_t>>& first,
-                  std::stack<std::uint64_t, std::vector<std::uint64_t>>& second);
+        bool swap(std::stack<std::uint_fast8_t, std::vector<std::uint_fast8_t>>& first,
+                  std::stack<std::uint_fast8_t, std::vector<std::uint_fast8_t>>& second);
 
     private:
         std::uint64_t numDisks;
 
-        std::stack<std::uint64_t, std::vector<std::uint64_t>> a;
-        std::stack<std::uint64_t, std::vector<std::uint64_t>> b;
-        std::stack<std::uint64_t, std::vector<std::uint64_t>> c;
+        std::stack<std::uint_fast8_t, std::vector<std::uint_fast8_t>> a;
+        std::stack<std::uint_fast8_t, std::vector<std::uint_fast8_t>> b;
+        std::stack<std::uint_fast8_t, std::vector<std::uint_fast8_t>> c;
 
         std::uint64_t numSteps;
 };
